+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 1. [](#new)
     * Added `creDate` page field to store date of page creation
+    * Added `lastAuthor` page field to store author of last page editing